Pin bijection-guava to jsr305 1.3.9.
The bijection-guava module works around guava only requiring jsr305 in
provided scope yet scalac needing all java annotations at compile time
by adding an explicit dep on jsr305.  It was using a valid ivy
revision string with a trailing +, ie: '1.3.+' - and this works fine
for sbt / ivy; however make-pom does not attempt to map ivy version
numbers to valid pom version numbers and a trailing + is treated as a
literal and not a ranged version by maven / pom consumers [1]. This
work-around just pins jsr305 to 1.3.9 which is what guava 14 asks for
(in provided scope) in its own pom.  Another valid approach is to use
'[1.3,)' as the version which is valid in ivy.xml, sbt and pom.xml,
but fwict the original use of '1.3.+' was just copypasta from this bug
report: http://code.google.com/p/guava-libraries/issues/detail?id=1095

[1] https://github.com/sbt/sbt/blob/b7c50d42698f8309c76ed5fd0efbb3ef61ee601e/ivy/src/main/scala/sbt/MakePom.scala